+++ src/test/kotlin/de/nordgedanken/auto_hotkey/lang/parser/AhkParsingTestBase.kt
@@ -30,15 +30,17 @@ abstract class AhkParsingTestBase(@NonNls dataPath: String) :
 
     protected fun hasError(file: PsiFile): Boolean {
         var hasErrors = false
-        file.accept(object : PsiElementVisitor() {
-            override fun visitElement(element: PsiElement) {
-                if (element is PsiErrorElement) {
-                    hasErrors = true
-                    return
+        file.accept(
+            object : PsiElementVisitor() {
+                override fun visitElement(element: PsiElement) {
+                    if (element is PsiErrorElement) {
+                        hasErrors = true
+                        return
+                    }
+                    element.acceptChildren(this)
                 }
-                element.acceptChildren(this)
             }
-        })
+        )
         return hasErrors
     }